feat: Add `web` command for scraping text from URLs
Introduced a new `web` command to the CLI that scrapes text from a specified URL. This command allows the user to either print the scraped text to the console or save it to a file. Handlers for both printing to the console and writing to a file have been implemented.